+++ R/rstudio.R
@@ -39,33 +39,17 @@ use_rstudio <- function(line_ending = c("posix", "windows")) {
 #' be found in the blog post [Project-oriented
 #' workflow](https://www.tidyverse.org/blog/2017/12/workflow-vs-script/).
 #'
-#' Only `use_blank_slate("project")` is automated so far, since RStudio
-#' currently only supports modification of user-level or global options via the
-#' user interface.
-#'
 #' @inheritParams edit
 #'
 #' @export
 use_blank_slate <- function(scope = c("user", "project")) {
   scope <- match.arg(scope)
 
-  if (scope == "user") { # nocov start
-    ui_todo(
-      "To start ALL RStudio sessions with a blank slate, \\
-      you must set this interactively, for now."
-    )
-    ui_todo(
-      "In {ui_field('Global Options > General')}, \\
-      do NOT check {ui_field('Restore .RData into workspace at startup')}."
-    )
-    ui_todo(
-      "In {ui_field('Global Options > General')}, \\
-      set {ui_field('Save workspace to .RData on exit')} to {ui_value('Never')}."
-    )
-    ui_todo(
-      "Call {ui_code('use_blank_slate(\"project\")')} to opt in to the \\
-      blank slate workflow for a specific project."
-    )
+  if (scope == "user") {
+    use_rstudio_config(list(
+      save_workspace = "never",
+      load_workspace = FALSE
+    ))
     return(invisible())
   } # nocov end
 
@@ -183,4 +167,36 @@ rstudio_config_path <- function(...) {
   base <- rappdirs::user_config_dir("RStudio", os = os)
 
   path(base, ...)
+}
+
+rstudio_prefs_read <- function() {
+  path <- rstudio_config_path("rstudio-prefs.json")
+  if (file_exists(path)) {
+    jsonlite::read_json(path)
+  } else {
+    list()
+  }
+}
+rstudio_prefs_write <- function(json) {
+  path <- rstudio_config_path("rstudio-prefs.json")
+  dir_create(path_dir(path))
+  jsonlite::write_json(json, path, auto_unbox = TRUE, pretty = TRUE)
+}
+
+use_rstudio_config <- function(values) {
+  stopifnot(is.list(values), is_named(values))
+  json <- rstudio_prefs_read()
+
+  for (name in names(values)) {
+    val <- values[[name]]
+
+    if (identical(json[[name]], val)) {
+      next
+    }
+
+    ui_done("Setting RStudio preference {ui_field(name)} to {ui_value(val)}")
+    json[[name]] <- val
+  }
+
+  rstudio_prefs_write(json)
 }