+++ NEWS.md
@@ -139,6 +139,8 @@ usethis is now fully cognizant of the [changes to RStudio preferences](https://b
 
 * New `edit_rstudio_prefs()` opens RStudio preferences file for editing (#1148).
 
+* `use_blank_slate()` now configures RStudio for you (#1018).
+
 ## Other changes
 
 * `browse_package()` and `browse_project()` are new functions that let the user choose from a list of URLs derived from local Git remotes and DESCRIPTION (local or possibly on CRAN) (#1113).